Completed Normal Layout assignment 1
I was asked to chnage the values of the box model, by altering the
exsisiting rule values of particualr selctors. One thing I learnt here
was on the auto value of the property margin. I was asked to push a
block as much right as possible, this can be acheived using margin-left
setting to auto.

Done - 01-margin-and-padding-1